+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - Install tree
 - Install composer
+- Install Java, Spring Boot, Maven and Gradle
+- Check if Docker is running before starting
 ### Changed
 - Change deprecated `brew cask reinstall` commands to just `brew reinstall`
 - `git-delete-all-local-branches-except-master-develop-and-current` becomes `git-delete-all-local-branches-except-main-develop-and-current`, but it still doesn't delete master.